+++ src/Conf.ml
@@ -871,6 +871,12 @@ module Formatting = struct
       (fun conf x -> {conf with match_indent_nested= x})
       (fun conf -> conf.match_indent_nested)
 
+  let default_max_indent =
+    (* Creating a fresh formatter in case the value of max-indent has been
+       changed for stdout. *)
+    let fs = Format.formatter_of_buffer (Buffer.create 0) in
+    Int.to_string (Format.pp_get_max_indent fs ())
+
   let max_indent =
     let docv = "COLS" in
     let doc =
@@ -878,7 +884,7 @@ module Formatting = struct
        the offset of the previous line."
     in
     C.any
-      Arg.(some int)
+      Arg.(some ~none:default_max_indent int)
       ~names:["max-indent"] ~doc ~docv ~section ~default:None
       ~allow_inline:false
       (fun conf x -> {conf with max_indent= x})